add helper function files
