+++ lib/components/Weather.js
@@ -1,4 +1,8 @@
 import React from 'react';
+import measureWeather from './helpers/measureweather';
+import transformScale from './helpers/transformscale';
+import transformWeatherType from './helpers/transformweathertype';
+
 
 const Weather = (props) => {
   let { location, date, weatherType, temp } = props
@@ -16,41 +20,5 @@ const Weather = (props) => {
     </div>
   )
 };
-
-function measureWeather(temp) {
-  let average = (temp.high+temp.low)/2
-  if (average > 75) {
-    return 'hot';
-  }
-  if (average <= 75 && average >= 51) {
-    return 'warm';
-  }
-  if (average <= 50 && average >= 26) {
-    return 'cool';
-  }
-  if (average <= 25) {
-    return 'cold';
-  }
-}
-
-function transformWeatherType(type) {
-  if (type === 'thunder storms') {
-    return 'thunderstorms'
-  } else if (type === 'cloudy') {
-    return 'clouds'
-  }  else if (type === 'sunny') {
-    return 'clear'
-  } else if (type === 'windy') {
-    return 'wind'
-  } else if (type === 'foggy') {
-    return 'fog'
-  } else { return type }
-}
-
-function transformScale(number) {
-  if (number === 3) {
-    return 'severe'
-  } else {return 'not-severe'}
-}
 
 export default Weather;
