+++ lib/components/helpers/transformscale.js
@@ -1,0 +1,7 @@
+function transformScale(number) {
+  if (number === 3) {
+    return 'severe'
+  } else {return 'not-severe'}
+}
+
+export default transformScale;
